requirements.txt
aiogram
apscheduler
python-dotenv 